Improve error message for non-sequence arguments to from_samples
